Refactor queries for composability
