+++ queries/ArticleBySlugQuery.ts
@@ -1,0 +1,15 @@
+import gql from 'graphql-tag';
+
+const ArticleBySlugQuery = gql`
+  query article($id: String!) {
+    articleBySlug(slug: $id) {
+      id
+      title
+      headerImageURL
+      summary
+      content
+    }
+  }
+`;
+
+export default ArticleBySlugQuery;
